third_party/blink/web_tests/external/wpt/scroll-animations/css/timeline-offset-keyframes-hidden-subject-expected.txt

This is a testharness.js-based test.
[FAIL] Keyframes with timeline-offsets ignored when timeline is inactive
  assert_equals: properties on ComputedKeyframe #0: Keyframes with invalid view timeline should match expected "composite,computedOffset,easing,offset,opacity" but got "composite,computedOffset,easing,marginRight,offset,opacity"
Harness: the test ran to completion.
